[#106][FE] canvasUtils로 함수 분리
